refactor(primitives): Extract module for SectorData
